Infra.Automation-Mistral/requirements.txt
urllib3==1.25.9
requests==2.24.0
PyYAML==5.3.1
python-jenkins==1.0.1
jenkinsapi

tabulate
beautifulsoup4